Move OpenStack scripts to openstack-lightspeed org
road-core/rag-content contains OpenStack specific scripts that
are needed if one wants to build a vector database for OpenStack
documentation.

This commit moves the scripts to the openstack-lightpseed/rag-content
repository which should from now on contain the OpenStack specific
code and road-core/rag-content should be only used as a provider
of common logic.

[1] https://github.com/road-core/rag-content