[GithubAction] Add build-and-push-rag-content workflow
This commit introduces a new workflow called build-and-push-rag-content.
This workflow builds a rag-content image which contains vector database
containing data from OpenStack documentation. Currently the workflow
only builds rag-content image for nova, neutron, cinder, keystone, glance
and swift projects. If the image build is successful then it is pushed
to quay.io/openstack-lightspeed/rag-content.

If the workflow got triggered by a PRs creation it only builds partial
image and does not perform a push of the image to the quay.io registry.